requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
websockets
pydantic
pytest